Fix slash in url and check for permissions
